+++ src/actions/authActions.js
@@ -10,6 +10,8 @@ import msgbox from '../helpers/msgbox'
 import log from '../helpers/log'
 import UrlManager from './urlManager'
 
+import { getUserRequests } from './repoActions'
+
 // object used to save the current location in the local storage
 // when the user pushes the login button.
 const urlManager = typeof window !== 'undefined' && new UrlManager(window)
@@ -29,7 +31,10 @@ export function start() {
         getProfile(token.id_token)
           .then(profile => {
             if (profile) {
-              return dispatch(loginSuccess(profile, token.id_token))
+              const action = dispatch(loginSuccess(profile, token.id_token))
+              console.info(profile.nickname)
+              return dispatch(getUserRequests(profile.nickname))
+              return action
             } else {
               return dispatch(loginFailure())
             }
@@ -52,6 +57,7 @@ export function login() {
   const redirect_uri = `${self.location.origin}%2Fauth0.html`
   const auth0Client = 'eyJuYW1lIjoiYXV0aDAuanMiLCJ2ZXJzaW9uIjoiNi44LjAifQ'
   const url = `${APP_URL}/authorize?scope=openid&response_type=token&connection=github&sso=true&client_id=${client_id}&redirect_uri=${redirect_uri}&auth0Client=${auth0Client}`
+  console.info(url)
   return dispatch => {
     dispatch(loginRequest())
     // Go to auth0 authenication page