+++ src/components/home/Home.jsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import { Link } from 'react-router'
 
 import log from '../../helpers/log'
 import MainContent from '../common/MainContent'
@@ -17,7 +18,17 @@ const addedSince = (hotFilter) => {
 const Home = React.createClass({
   render() {
     log('Render the <Home> component', this.props)
-    const { hotProjects, popularProjects, maxStars, isLoggedin, uiActions, hotFilter, showMetrics } = this.props
+    const {
+      hotProjects,
+      popularProjects,
+      maxStars,
+      isLoggedin,
+      pending,
+      uiActions,
+      hotFilter,
+      showMetrics,
+      authActions
+    } = this.props
     return (
       <MainContent>
         { this.props.errorMessage && <ErrorMessage text={ this.props.errorMessage } /> }
@@ -96,9 +107,54 @@ const Home = React.createClass({
             </div>
           </div>
         )}
+        <MoreProjects
+          handleClick={authActions.login}
+          isLoggedin={isLoggedin}
+          pending={pending}
+        />
       </MainContent>
     )
   }
 })
 
-module.exports = Home
+const MoreProjects1 = ({ handleClick, isLoggedin }) => (
+  <div className="card">
+    <div className="header">Do you want more projects ?</div>
+    {isLoggedin ? (
+      <Link className="inner link" to="/requests/add-project">
+        Add a project on Github
+      </Link>
+    ) : (
+      <div className="inner link" onClick={handleClick}>
+        <span className="octicon octicon-mark-github" />
+        {' '}
+        Sign in with Github to add a new project.
+      </div>
+    )}
+  </div>
+)
+const MoreProjects = ({ handleClick, isLoggedin, pending }) => {
+  return (
+    <div>
+      <h3 className="with-comment">Do you want more projects ?</h3>
+      {isLoggedin ? (
+        <Link className="btn block button-outline" to="/requests/add-project">
+          <span className="octicon octicon-plus" />
+          {' '}
+          Add a project on Github
+        </Link>
+      ) : (
+        <button
+          className={`btn block button-outline${pending ? ' ui loading button' : ''}`}
+          onClick={handleClick}
+        >
+          <span className="octicon octicon-mark-github" />
+          {' '}
+          Sign in with Github to add a new project
+        </button>
+      )}
+    </div>
+  )
+}
+
+export default Home